Implement Polls endpoints
Implemented GET and POST /api/polls.  GET is a protected endpoint.  POST
is unprotected, but checks if the user is logged in and, if so,
associates the new poll with the user's ID.  Also implemented GET,
PATCH, and DELETE /api/polls/:id.  GET is unprotected, PATCH and DELETE
are protected.